refactor: エラーハンドリングの配置ディレクトリ
